Test.Goblin.Core

Description

The core typeclasses and associated methods of goblins.

Documentation

data GoblinData g

The state we carry as we perform goblins actions.

Constructors

GoblinData

Fields

_genes ∷ !(Genome g)

Remaining genes, controlling how a goblin operates.

_bagOfTricks ∷ !(TypeRepMap [])

A goblin's bag of tricks contains items of many differnt types. When tinkering, a goblin (depending on its genome) might look in its bag of tricks to see whether it has anything of the appropriate type to replace what it's currently tinkering with (or, depending on the type, do something different - for example, utilise a monoid instance to add things together).

class SeedGoblin a where

Recur down a datatype, adding the sub-datatypes to the TinkerM TypeRepMap

Minimal complete definition

Nothing

Methods

seeder ∷ a → TinkerM g ()

Recur down a type, adding elements to the TypeRepMap

seeder ∷ Typeable a ⇒ a → TinkerM g ()

Recur down a type, adding elements to the TypeRepMap

class AddShrinks a where

Whereas pure creates a Hedgehog tree with no shrinks, addShrinks creates a tree with shrinks.

Minimal complete definition

Nothing

Methods

addShrinks ∷ a → Gen a

addShrinks ∷ Enum a ⇒ a → Gen a

class GeneOps g where

A typeclass for actions over genomes.

Methods

onGene

Arguments

∷ TinkerM g a
When gene is on.

→ TinkerM g a
When gene is off.

→ TinkerM g a

Choose between two actions based on the value of a gene.

transcribeGenesAsInt ∷ Int → TinkerM g Int

Transcribe sufficient genes to get an integer in the range [0..n].

class (GeneOps g, Typeable a) ⇒ Goblin g a where

The interface to goblins. This class defines two actions - tinkering with an existing value - conjureing a new value

Methods

tinker ∷ Gen a → TinkerM g (Gen a)

Tinker with an item of type a.

conjure ∷ TinkerM g a

As well as tinkering, goblins can conjure fresh items into existence.

type TinkerM g = State (GoblinData g)

Tinker monad.

bagOfTricks ∷ ∀ g. Lens' (GoblinData g) (TypeRepMap [])

genes ∷ ∀ g g. Lens (GoblinData g) (GoblinData g) (Genome g) (Genome g)

saveInBagOfTricks ∷ ∀ g a. Typeable a ⇒ a → TinkerM g a

Helper function to save a value in the bagOfTricks, and return it.

tinkerWithToys ∷ (AddShrinks a, Goblin g a) ⇒ [Gen a → Gen a → Gen a] → Gen a → TinkerM g (Gen a)

Construct a tinker function given a set of possible things to do.

Each toy is a function taking the original value and one grabbed from the bag of tricks or conjured.

tinkerRummagedOrConjureOrSave ∷ (Goblin g a, AddShrinks a) ⇒ TinkerM g (Gen a) → TinkerM g (Gen a)

Either tinker with a rummaged value, conjure a new value, or save the argument in the bagOfTricks and return it.

transcribeGene ∷ TinkerM g g

Read (and consume) a gene from the genome.

rummage ∷ ∀ a g. (GeneOps g, Typeable a) ⇒ TinkerM g (Maybe a)

Fetch something from the bag of tricks if there's something there.

rummageAll ∷ ∀ a g. Typeable a ⇒ TinkerM g [a]

Fetch everything from the bag of tricks.

rummageOrConjure ∷ ∀ a g. Goblin g a ⇒ TinkerM g a

Fetch something from the bag of tricks, or else conjure it up.

tinkerRummagedOrConjure ∷ ∀ a g. (Goblin g a, AddShrinks a) ⇒ TinkerM g (Gen a)

Attempt to rummage. If a value is available, either tinker with it or leave it intact. If no value is available, conjure a fresh one and add shrinks to it.

shrinkEnum ∷ Enum a ⇒ a → [a]

Use an Enum instance to create a shrink tree which shrinks towards `toEnum 0`.

mkGoblin ∷ Genome g → TypeRepMap [] → GoblinData g

Spawn a goblin from a given genome and a bag of tricks.

mkEmptyGoblin ∷ Genome g → GoblinData g

Spawn a goblin from a genome, with an empty TypeRepMap.

geneListIndex ∷ GeneOps g ⇒ [a] → TinkerM g Int

Use the genome to generate an index within the bounds of the provided list.

genPopulation ∷ Gen (Population Bool)

Convenience Hedgehog generator.

(<$$>) ∷ (Functor f, Functor g) ⇒ (a → b) → f (g a) → f (g b)

<$> through nested functors.

(<**>) ∷ (Applicative f, Applicative g) ⇒ f (g (a → b)) → f (g a) → f (g b)

<*> through nested functors.