Test.Goblin.Instances

Orphan instances

SeedGoblin Bool

Instance details

Methods

seeder ∷ Bool → TinkerM g ()

SeedGoblin Char

Instance details

Methods

seeder ∷ Char → TinkerM g ()

SeedGoblin Double

Instance details

Methods

seeder ∷ Double → TinkerM g ()

SeedGoblin Int

Instance details

Methods

seeder ∷ Int → TinkerM g ()

SeedGoblin Integer

Instance details

Methods

seeder ∷ Integer → TinkerM g ()

SeedGoblin Natural

Instance details

Methods

seeder ∷ Natural → TinkerM g ()

SeedGoblin Word8

Instance details

Methods

seeder ∷ Word8 → TinkerM g ()

SeedGoblin Word64

Instance details

Methods

seeder ∷ Word64 → TinkerM g ()

SeedGoblin ()

Instance details

Methods

seeder ∷ () → TinkerM g ()

AddShrinks Bool

Instance details

Methods

addShrinks ∷ Bool → Gen Bool

AddShrinks Char

Instance details

Methods

addShrinks ∷ Char → Gen Char

AddShrinks Double

Instance details

Methods

addShrinks ∷ Double → Gen Double

AddShrinks Int

Instance details

Methods

addShrinks ∷ Int → Gen Int

AddShrinks Integer

Instance details

Methods

addShrinks ∷ Integer → Gen Integer

AddShrinks Natural

Instance details

Methods

addShrinks ∷ Natural → Gen Natural

AddShrinks Word8

Instance details

Methods

addShrinks ∷ Word8 → Gen Word8

AddShrinks Word64

Instance details

Methods

addShrinks ∷ Word64 → Gen Word64

AddShrinks ()

Instance details

Methods

addShrinks ∷ () → Gen ()

GeneOps Bool

Instance details

Methods

onGene ∷ TinkerM Bool a → TinkerM Bool a → TinkerM Bool a

transcribeGenesAsInt ∷ Int → TinkerM Bool Int

GeneOps a ⇒ Goblin a Double
This instance generates Double values in range 0..1 at 0.01 increments. 0.01, 0.02 ... 0.99, 1.00

Instance details

Methods

tinker ∷ Gen Double → TinkerM a (Gen Double)

conjure ∷ TinkerM a Double

GeneOps a ⇒ Goblin a Word64

Instance details

Methods

tinker ∷ Gen Word64 → TinkerM a (Gen Word64)

conjure ∷ TinkerM a Word64

GeneOps a ⇒ Goblin a Int

Instance details

Methods

tinker ∷ Gen Int → TinkerM a (Gen Int)

conjure ∷ TinkerM a Int

GeneOps a ⇒ Goblin a Natural

Instance details

Methods

tinker ∷ Gen Natural → TinkerM a (Gen Natural)

conjure ∷ TinkerM a Natural

GeneOps a ⇒ Goblin a Integer

Instance details

Methods

tinker ∷ Gen Integer → TinkerM a (Gen Integer)

conjure ∷ TinkerM a Integer

GeneOps a ⇒ Goblin a Char

Instance details

Methods

tinker ∷ Gen Char → TinkerM a (Gen Char)

conjure ∷ TinkerM a Char

GeneOps a ⇒ Goblin a Bool

Instance details

Methods

tinker ∷ Gen Bool → TinkerM a (Gen Bool)

conjure ∷ TinkerM a Bool

(Goblin genome arg, AddShrinks arg) ⇒ Goblin genome (Ratio arg)

Instance details

Methods

tinker ∷ Gen (Ratio arg) → TinkerM genome (Gen (Ratio arg))

conjure ∷ TinkerM genome (Ratio arg)

(Goblin g a, Ord a, AddShrinks a, Typeable a) ⇒ Goblin g (Set a)

Instance details

Methods

tinker ∷ Gen (Set a) → TinkerM g (Gen (Set a))

conjure ∷ TinkerM g (Set a)

(AddShrinks a, Eq a, Typeable a, GeneOps g, Goblin g a) ⇒ Goblin g [a]
Our list goblin behaves slightly differently, since it pulls whole lists of things from the bag of tricks, and is also specialised to do some more messing about with lists.

Instance details

Methods

tinker ∷ Gen [a] → TinkerM g (Gen [a])

conjure ∷ TinkerM g [a]

(Goblin g a, AddShrinks a) ⇒ Goblin g (Maybe a)

Instance details

Methods

tinker ∷ Gen (Maybe a) → TinkerM g (Gen (Maybe a))

conjure ∷ TinkerM g (Maybe a)

(Goblin genome arg1, AddShrinks arg1, Goblin genome arg2, AddShrinks arg2) ⇒ Goblin genome (arg1, arg2)

Instance details

Methods

tinker ∷ Gen (arg1, arg2) → TinkerM genome (Gen (arg1, arg2))

conjure ∷ TinkerM genome (arg1, arg2)

(Goblin g k, Goblin g v, Ord k, Eq k, Eq v, AddShrinks (Map k v), AddShrinks k, AddShrinks v, Typeable k, Typeable v) ⇒ Goblin g (Map k v)

Instance details

Methods

tinker ∷ Gen (Map k v) → TinkerM g (Gen (Map k v))

conjure ∷ TinkerM g (Map k v)

(Goblin genome arg1, AddShrinks arg1, Goblin genome arg2, AddShrinks arg2, Goblin genome arg3, AddShrinks arg3) ⇒ Goblin genome (arg1, arg2, arg3)

Instance details

Methods

tinker ∷ Gen (arg1, arg2, arg3) → TinkerM genome (Gen (arg1, arg2, arg3))

conjure ∷ TinkerM genome (arg1, arg2, arg3)

(Goblin genome arg1, AddShrinks arg1, Goblin genome arg2, AddShrinks arg2, Goblin genome arg3, AddShrinks arg3, Goblin genome arg4, AddShrinks arg4) ⇒ Goblin genome (arg1, arg2, arg3, arg4)

Instance details

Methods

tinker ∷ Gen (arg1, arg2, arg3, arg4) → TinkerM genome (Gen (arg1, arg2, arg3, arg4))

conjure ∷ TinkerM genome (arg1, arg2, arg3, arg4)

(Goblin genome arg1, AddShrinks arg1, Goblin genome arg2, AddShrinks arg2, Goblin genome arg3, AddShrinks arg3, Goblin genome arg4, AddShrinks arg4, Goblin genome arg5, AddShrinks arg5) ⇒ Goblin genome (arg1, arg2, arg3, arg4, arg5)

Instance details

Methods

tinker ∷ Gen (arg1, arg2, arg3, arg4, arg5) → TinkerM genome (Gen (arg1, arg2, arg3, arg4, arg5))

conjure ∷ TinkerM genome (arg1, arg2, arg3, arg4, arg5)

(Goblin genome arg1, AddShrinks arg1, Goblin genome arg2, AddShrinks arg2, Goblin genome arg3, AddShrinks arg3, Goblin genome arg4, AddShrinks arg4, Goblin genome arg5, AddShrinks arg5, Goblin genome arg6, AddShrinks arg6) ⇒ Goblin genome (arg1, arg2, arg3, arg4, arg5, arg6)

Instance details

Methods

tinker ∷ Gen (arg1, arg2, arg3, arg4, arg5, arg6) → TinkerM genome (Gen (arg1, arg2, arg3, arg4, arg5, arg6))

conjure ∷ TinkerM genome (arg1, arg2, arg3, arg4, arg5, arg6)

(Goblin genome arg1, AddShrinks arg1, Goblin genome arg2, AddShrinks arg2, Goblin genome arg3, AddShrinks arg3, Goblin genome arg4, AddShrinks arg4, Goblin genome arg5, AddShrinks arg5, Goblin genome arg6, AddShrinks arg6, Goblin genome arg7, AddShrinks arg7) ⇒ Goblin genome (arg1, arg2, arg3, arg4, arg5, arg6, arg7)

Instance details

Methods

tinker ∷ Gen (arg1, arg2, arg3, arg4, arg5, arg6, arg7) → TinkerM genome (Gen (arg1, arg2, arg3, arg4, arg5, arg6, arg7))

conjure ∷ TinkerM genome (arg1, arg2, arg3, arg4, arg5, arg6, arg7)

(Goblin genome arg1, AddShrinks arg1, Goblin genome arg2, AddShrinks arg2, Goblin genome arg3, AddShrinks arg3, Goblin genome arg4, AddShrinks arg4, Goblin genome arg5, AddShrinks arg5, Goblin genome arg6, AddShrinks arg6, Goblin genome arg7, AddShrinks arg7, Goblin genome arg8, AddShrinks arg8) ⇒ Goblin genome (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8)

Instance details

Methods

tinker ∷ Gen (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8) → TinkerM genome (Gen (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8))

conjure ∷ TinkerM genome (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8)

(Goblin genome arg1, AddShrinks arg1, Goblin genome arg2, AddShrinks arg2, Goblin genome arg3, AddShrinks arg3, Goblin genome arg4, AddShrinks arg4, Goblin genome arg5, AddShrinks arg5, Goblin genome arg6, AddShrinks arg6, Goblin genome arg7, AddShrinks arg7, Goblin genome arg8, AddShrinks arg8, Goblin genome arg9, AddShrinks arg9) ⇒ Goblin genome (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9)

Instance details

Methods

tinker ∷ Gen (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9) → TinkerM genome (Gen (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9))

conjure ∷ TinkerM genome (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9)

(Goblin genome arg1, AddShrinks arg1, Goblin genome arg2, AddShrinks arg2, Goblin genome arg3, AddShrinks arg3, Goblin genome arg4, AddShrinks arg4, Goblin genome arg5, AddShrinks arg5, Goblin genome arg6, AddShrinks arg6, Goblin genome arg7, AddShrinks arg7, Goblin genome arg8, AddShrinks arg8, Goblin genome arg9, AddShrinks arg9, Goblin genome arg10, AddShrinks arg10) ⇒ Goblin genome (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10)

Instance details

Methods

tinker ∷ Gen (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10) → TinkerM genome (Gen (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10))

conjure ∷ TinkerM genome (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10)

(Goblin genome arg1, AddShrinks arg1, Goblin genome arg2, AddShrinks arg2, Goblin genome arg3, AddShrinks arg3, Goblin genome arg4, AddShrinks arg4, Goblin genome arg5, AddShrinks arg5, Goblin genome arg6, AddShrinks arg6, Goblin genome arg7, AddShrinks arg7, Goblin genome arg8, AddShrinks arg8, Goblin genome arg9, AddShrinks arg9, Goblin genome arg10, AddShrinks arg10, Goblin genome arg11, AddShrinks arg11) ⇒ Goblin genome (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10, arg11)

Instance details

Methods

tinker ∷ Gen (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10, arg11) → TinkerM genome (Gen (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10, arg11))

conjure ∷ TinkerM genome (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10, arg11)

(SeedGoblin a, Typeable a) ⇒ SeedGoblin [a]

Instance details

Methods

seeder ∷ [a] → TinkerM g ()

(SeedGoblin a, Typeable a) ⇒ SeedGoblin (Maybe a)

Instance details

Methods

seeder ∷ Maybe a → TinkerM g ()

(SeedGoblin a, Typeable a) ⇒ SeedGoblin (Seq a)

Instance details

Methods

seeder ∷ Seq a → TinkerM g ()

(SeedGoblin a, Typeable a) ⇒ SeedGoblin (Set a)

Instance details

Methods

seeder ∷ Set a → TinkerM g ()

AddShrinks a ⇒ AddShrinks [a]

Instance details

Methods

addShrinks ∷ [a] → Gen [a]

AddShrinks a ⇒ AddShrinks (Maybe a)

Instance details

Methods

addShrinks ∷ Maybe a → Gen (Maybe a)

AddShrinks arg ⇒ AddShrinks (Ratio arg)

Instance details

Methods

addShrinks ∷ Ratio arg → Gen (Ratio arg)

(AddShrinks a, Ord a) ⇒ AddShrinks (Set a)

Instance details

Methods

addShrinks ∷ Set a → Gen (Set a)

(SeedGoblin arg1, Typeable arg1, SeedGoblin arg2, Typeable arg2) ⇒ SeedGoblin (arg1, arg2)

Instance details

Methods

seeder ∷ (arg1, arg2) → TinkerM g ()

(SeedGoblin a, Typeable a, SeedGoblin b, Typeable b) ⇒ SeedGoblin (Map a b)

Instance details

Methods

seeder ∷ Map a b → TinkerM g ()

(SeedGoblin a, Typeable a, SeedGoblin b, Typeable b) ⇒ SeedGoblin (Bimap a b)

Instance details

Methods

seeder ∷ Bimap a b → TinkerM g ()

(AddShrinks arg1, AddShrinks arg2) ⇒ AddShrinks (arg1, arg2)

Instance details

Methods

addShrinks ∷ (arg1, arg2) → Gen (arg1, arg2)

(AddShrinks k, Ord k, AddShrinks v) ⇒ AddShrinks (Map k v)

Instance details

Methods

addShrinks ∷ Map k v → Gen (Map k v)

(SeedGoblin arg1, Typeable arg1, SeedGoblin arg2, Typeable arg2, SeedGoblin arg3, Typeable arg3) ⇒ SeedGoblin (arg1, arg2, arg3)

Instance details

Methods

seeder ∷ (arg1, arg2, arg3) → TinkerM g ()

(AddShrinks arg1, AddShrinks arg2, AddShrinks arg3) ⇒ AddShrinks (arg1, arg2, arg3)

Instance details

Methods

addShrinks ∷ (arg1, arg2, arg3) → Gen (arg1, arg2, arg3)

(SeedGoblin arg1, Typeable arg1, SeedGoblin arg2, Typeable arg2, SeedGoblin arg3, Typeable arg3, SeedGoblin arg4, Typeable arg4) ⇒ SeedGoblin (arg1, arg2, arg3, arg4)

Instance details

Methods

seeder ∷ (arg1, arg2, arg3, arg4) → TinkerM g ()

(AddShrinks arg1, AddShrinks arg2, AddShrinks arg3, AddShrinks arg4) ⇒ AddShrinks (arg1, arg2, arg3, arg4)

Instance details

Methods

addShrinks ∷ (arg1, arg2, arg3, arg4) → Gen (arg1, arg2, arg3, arg4)

(SeedGoblin arg1, Typeable arg1, SeedGoblin arg2, Typeable arg2, SeedGoblin arg3, Typeable arg3, SeedGoblin arg4, Typeable arg4, SeedGoblin arg5, Typeable arg5) ⇒ SeedGoblin (arg1, arg2, arg3, arg4, arg5)

Instance details

Methods

seeder ∷ (arg1, arg2, arg3, arg4, arg5) → TinkerM g ()

(AddShrinks arg1, AddShrinks arg2, AddShrinks arg3, AddShrinks arg4, AddShrinks arg5) ⇒ AddShrinks (arg1, arg2, arg3, arg4, arg5)

Instance details

Methods

addShrinks ∷ (arg1, arg2, arg3, arg4, arg5) → Gen (arg1, arg2, arg3, arg4, arg5)

(SeedGoblin arg1, Typeable arg1, SeedGoblin arg2, Typeable arg2, SeedGoblin arg3, Typeable arg3, SeedGoblin arg4, Typeable arg4, SeedGoblin arg5, Typeable arg5, SeedGoblin arg6, Typeable arg6) ⇒ SeedGoblin (arg1, arg2, arg3, arg4, arg5, arg6)

Instance details

Methods

seeder ∷ (arg1, arg2, arg3, arg4, arg5, arg6) → TinkerM g ()

(AddShrinks arg1, AddShrinks arg2, AddShrinks arg3, AddShrinks arg4, AddShrinks arg5, AddShrinks arg6) ⇒ AddShrinks (arg1, arg2, arg3, arg4, arg5, arg6)

Instance details

Methods

addShrinks ∷ (arg1, arg2, arg3, arg4, arg5, arg6) → Gen (arg1, arg2, arg3, arg4, arg5, arg6)

(SeedGoblin arg1, Typeable arg1, SeedGoblin arg2, Typeable arg2, SeedGoblin arg3, Typeable arg3, SeedGoblin arg4, Typeable arg4, SeedGoblin arg5, Typeable arg5, SeedGoblin arg6, Typeable arg6, SeedGoblin arg7, Typeable arg7) ⇒ SeedGoblin (arg1, arg2, arg3, arg4, arg5, arg6, arg7)

Instance details

Methods

seeder ∷ (arg1, arg2, arg3, arg4, arg5, arg6, arg7) → TinkerM g ()

(AddShrinks arg1, AddShrinks arg2, AddShrinks arg3, AddShrinks arg4, AddShrinks arg5, AddShrinks arg6, AddShrinks arg7) ⇒ AddShrinks (arg1, arg2, arg3, arg4, arg5, arg6, arg7)

Instance details

Methods

addShrinks ∷ (arg1, arg2, arg3, arg4, arg5, arg6, arg7) → Gen (arg1, arg2, arg3, arg4, arg5, arg6, arg7)

(SeedGoblin arg1, Typeable arg1, SeedGoblin arg2, Typeable arg2, SeedGoblin arg3, Typeable arg3, SeedGoblin arg4, Typeable arg4, SeedGoblin arg5, Typeable arg5, SeedGoblin arg6, Typeable arg6, SeedGoblin arg7, Typeable arg7, SeedGoblin arg8, Typeable arg8) ⇒ SeedGoblin (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8)

Instance details

Methods

seeder ∷ (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8) → TinkerM g ()

(AddShrinks arg1, AddShrinks arg2, AddShrinks arg3, AddShrinks arg4, AddShrinks arg5, AddShrinks arg6, AddShrinks arg7, AddShrinks arg8) ⇒ AddShrinks (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8)

Instance details

Methods

addShrinks ∷ (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8) → Gen (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8)

(SeedGoblin arg1, Typeable arg1, SeedGoblin arg2, Typeable arg2, SeedGoblin arg3, Typeable arg3, SeedGoblin arg4, Typeable arg4, SeedGoblin arg5, Typeable arg5, SeedGoblin arg6, Typeable arg6, SeedGoblin arg7, Typeable arg7, SeedGoblin arg8, Typeable arg8, SeedGoblin arg9, Typeable arg9) ⇒ SeedGoblin (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9)

Instance details

Methods

seeder ∷ (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9) → TinkerM g ()

(AddShrinks arg1, AddShrinks arg2, AddShrinks arg3, AddShrinks arg4, AddShrinks arg5, AddShrinks arg6, AddShrinks arg7, AddShrinks arg8, AddShrinks arg9) ⇒ AddShrinks (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9)

Instance details

Methods

addShrinks ∷ (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9) → Gen (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9)

(SeedGoblin arg1, Typeable arg1, SeedGoblin arg2, Typeable arg2, SeedGoblin arg3, Typeable arg3, SeedGoblin arg4, Typeable arg4, SeedGoblin arg5, Typeable arg5, SeedGoblin arg6, Typeable arg6, SeedGoblin arg7, Typeable arg7, SeedGoblin arg8, Typeable arg8, SeedGoblin arg9, Typeable arg9, SeedGoblin arg10, Typeable arg10) ⇒ SeedGoblin (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10)

Instance details

Methods

seeder ∷ (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10) → TinkerM g ()

(AddShrinks arg1, AddShrinks arg2, AddShrinks arg3, AddShrinks arg4, AddShrinks arg5, AddShrinks arg6, AddShrinks arg7, AddShrinks arg8, AddShrinks arg9, AddShrinks arg10) ⇒ AddShrinks (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10)

Instance details

Methods

addShrinks ∷ (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10) → Gen (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10)

(SeedGoblin arg1, Typeable arg1, SeedGoblin arg2, Typeable arg2, SeedGoblin arg3, Typeable arg3, SeedGoblin arg4, Typeable arg4, SeedGoblin arg5, Typeable arg5, SeedGoblin arg6, Typeable arg6, SeedGoblin arg7, Typeable arg7, SeedGoblin arg8, Typeable arg8, SeedGoblin arg9, Typeable arg9, SeedGoblin arg10, Typeable arg10, SeedGoblin arg11, Typeable arg11) ⇒ SeedGoblin (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10, arg11)

Instance details

Methods

seeder ∷ (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10, arg11) → TinkerM g ()

(AddShrinks arg1, AddShrinks arg2, AddShrinks arg3, AddShrinks arg4, AddShrinks arg5, AddShrinks arg6, AddShrinks arg7, AddShrinks arg8, AddShrinks arg9, AddShrinks arg10, AddShrinks arg11) ⇒ AddShrinks (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10, arg11)

Instance details

Methods

addShrinks ∷ (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10, arg11) → Gen (arg1, arg2, arg3, arg4, arg5, arg6, arg7, arg8, arg9, arg10, arg11)